// packages/data-presenter/src/types/MESSAGE_TYPES.ts
export enum MESSAGE_TYPES {
  SIGN_IN = 'SIGN_IN',
  NETWORK_CALL = 'NETWORK_CALL',
  REQUEST_USER = 'REQUEST_USER',
  SEND_USER = 'SEND_USER',
  SYNC_REQUESTS = 'SYNC_REQUESTS',
  REQUEST_USAGE = 'REQUEST_USAGE',
  REQUEST_CREDENTIALS = 'REQUEST_CREDENTIALS',
  SEND_CREDENTIALS = 'SEND_CREDENTIALS',
  REQUEST_RESET_COUNTER = 'REQUEST_RESET_COUNTER',
  REQUEST_BLACKLISTED_PAGES = 'REQUEST_BLACKLISTED_PAGES',
  SEND_BLACKLISTED_PAGES = 'SEND_BLACKLISTED_PAGES',
  ADD_BLACKLISTED_PAGE = 'ADD_BLACKLISTED_PAGE',
  DELETE_BLACKLISTED_PAGE = 'DELETE_BLACKLISTED_PAGE',
  REQUEST_REPORTS = 'REQUEST_REPORTS',
  SEND_REPORTS = 'SEND_REPORTS',
  REQUEST_USING_MONITOR = 'REQUEST_USING_MONITOR',
  SEND_USING_MONITOR = 'SEND_USING_MONITOR',
  SET_USING_MONITOR = 'SET_USING_MONITOR',
  REQUEST_TIPS = 'REQUEST_TIPS',
  SEND_TIPS = 'SEND_TIPS',
  REQUEST_EXTENDED_POLLUTION = 'REQUEST_EXTENDED_POLLUTION',
  SEND_EXTENDED_POLLUTION = 'SEND_EXTENDED_POLLUTION',
  SET_EXTENDED_POLLUTION = 'SET_EXTENDED_POLLUTION'
}